feat: Added song search from SongSearch project
I previously worked on a project called SongSearch, which is also
a repo on my GitHub. The purpose of this project was to be able
to search for songs, albums, or artist using the Spotify web API.
This is functionality that should definitely be included in
JammyJamz as well. Therefore, the necessary classes have been
imported into this project. In addition, when the user clicks
"Ok" within the post type dialog, PostSearchResults is now
launched, which was the main activity of the SongSearch project.