feat: Saved access token to SharedPreferences
There were some serious issues with the way that I was obtaining
the access token in my previous commits - if PostSearch was
launched more than once, then the activity would try to obtain
the access token again. Spotify would then deny the subsequent
requests. I'm not sure of the exact reasoning for that, but I'm
sure that it has something to do with limiting the number of
times requests are made for access tokens.

In order to get around that, I now save the access token to
SharedPreferences within PostSearch and extract the access
token from there as well. I still need to detect when the
access token is expired and another request needs to be made...
That needs to be taken care of soon.