feat: User can now create a post
A user can now add a post to the Newsfeed activity. This required
quite a few adjustments, such as adding a send button to
WritePost, adjusting the ViewHolder layout of the Newsfeed,
adding a new member variable to my Post class, and finally having
Newsfeed handle what happens when it receives a new Post.