feat: Adjusted layout of post item
The layout for the list items within the newsfeed was bothering
me so much - not very clean looking at all. Spent some time
adjusting this. I now use the Palette class to add a nice
background behind the album artwork and adjust the title and
artist font colors to match the new background (while keeping
them legible). Also added a profile picture next to the
username. There have been a couple of other small UI changes,
such as adjusting the text size of some of the elements within
the list item.