feat: Added a "Jam!" button to WritePost
Added a button within the toolbar of WritePost that actually adds
the post and removed the send icon that was below the EditText.
The main reason I switched the placement of this element is
because I noticed many high class apps (Facebook, LinkedIn, etc.)
place this action item within their toolbar.