feat: Added a preview of post in WritePost; general clean-up
Added some views to the WritePost layout to include the username,
(eventual) profile picture, and the album art + background
to be shown when the user writes their post. Also created a
Utilities class for general functionality, such as adding
getter and setter methods for the username since these could be
useful across the app.