feat: Adjusted TextViews in post and WritePost
The TextViews in the list item layout for posts within the
newsfeed as well as within WritePost (for the post preview)
were sort of all over the place with their dimensions. They would
cover the album artwork if the text was too long, which obviously
had to be fixed. I've adjusted the layouts so that this no
longer happens.