feat: Search type dialog appears when widget button is clicked
Made some minor adjustments to the structure of PostSearch so
that it now has another extra. This extra is used to determine
if the PostTypeSelection dialog should be shown or not. If the
user clicks on the button in the widget, the dialog is shown
(since there really is no way for them to select what type of
post they'd like to add within the widget itself). However,
if the FAB is clicked on within Newsfeed, then the dialog
won't be shown in PostSearch.