feat: Added offline capabilities
Did quite a bit in this one. First, I extended the Application
class so that I can setup some global functionalities (things
that will be enabled regardless of what activity is displayed).
Within the JammyJamzApplication class, I enabled data persistence
within onCreate(). Also, within all of the activities, I added
a TextView which indicates to the user if they've lost
connection to the Firebase Realtime Database. This involved
some slight adjustments to the layout XMLs. Then, in the
application class, I added a helper function to either show
or hide the TextView, depending on the connection state.