feat: Reversed the order of displayed content in Newsfeed
For a modern newsfeed, the most recent items are always
displayed first to the end user with older content shown as the
user scrolls down. In all previous versions of the app, the
most recent posts were always shown at the end. I've adjusted
the RecyclerView to show the most recent posts at the top of
the newsfeed.