feat: Worked on ListView within widget
Added a ListView within the widget so that it now displays
relevant app data to the end user. However, it doesn't yet
display any artwork within the ListView since this apparently
becomes complicated when trying to use Picasso. I will need
to investigate this some more in order to get this to work.