feat: Created an IntentService to search for tracks
The Udacity rubric requires us to use several different means of
performing async processes, none of which I was using initially
for the project. The easiest place to make adjustments to meet
this requirement was the PostSearch activity since it definitely
uses async procs to make search requests.

I've adjusted PostSearch so that all search requests are sent
to SearchService, which is an IntentService. The service then
sends a local broadcast back to PostSearch which updates the
RecyclerView with the search results. As of now, it only works
for tracks and it uses Serializable instead of Parcelable.
Serious adjustments are still needed.