Removed DB listener in application class
Back to the swing of things on this one. I had a DatabaseReference
object in my application class that I used to display or not
display TextView for no Internet connection. I removed this since
there could be a lot of problems this this - our application
should definitely not have anything to do with updating views
within different activities. For the time being, I've moved this
logic within each activity, but I'd like to clean this up
further hopefully in the future.